corrected invalid read destructor pipes
