+++ src/analyzer/analyzer.c
@@ -6,7 +6,7 @@
 /*   By: echavez- <echavez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/05 15:04:10 by echavez-          #+#    #+#             */
-/*   Updated: 2023/10/10 17:27:40 by echavez-         ###   ########.fr       */
+/*   Updated: 2023/10/13 17:20:25 by echavez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,40 +24,4 @@ int	ft_analyzer(t_sh *sh)
 		return (EXIT_FAILURE);
 	ft_syntax_tree(sh);
 	return (EXIT_SUCCESS);
-}
-
-static void	free_ast(t_ast **ast)
-{
-	t_ast	*tmp;
-
-	while (*ast)
-	{
-		tmp = *ast;
-		*ast = (*ast)->next;
-		free(tmp->av);
-		free(tmp->redir);
-		free(tmp);
-	}
-	*ast = NULL;
-}
-
-void	analyzer_destructor(t_sh *sh)
-{
-	int	i;
-
-	if (sh->cl.tokens)
-		ft_free_split(&sh->cl.tokens);
-	if (sh->cl.cmds)
-	{
-		i = 0;
-		while (sh->cl.cmds[i])
-		{
-			free(sh->cl.cmds[i]);
-			i++;
-		}
-		free(sh->cl.cmds);
-		sh->cl.cmds = NULL;
-	}
-	if (sh->cl.ast)
-		free_ast(&sh->cl.ast);
 }
