+++ src/evaluator/which.c
@@ -6,7 +6,7 @@
 /*   By: echavez- <echavez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 17:51:33 by echavez-          #+#    #+#             */
-/*   Updated: 2023/10/13 17:14:29 by echavez-         ###   ########.fr       */
+/*   Updated: 2023/10/13 17:16:23 by echavez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
