+++ src/redirections/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: echavez- <echavez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/27 19:33:32 by echavez-          #+#    #+#             */
-/*   Updated: 2023/10/13 16:09:07 by echavez-         ###   ########.fr       */
+/*   Updated: 2023/10/13 17:30:06 by echavez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,15 +36,9 @@ static void	heredoc_prompt(char *delimiter, int *pipe)
 
 static void	execute_in_context(int *pipe_fd, char *delimeter)
 {
-	int	sig_bkp;
-
-	sig_bkp = g_sigint;
-	g_sigint = 0;
 	ft_signals(HEREDOC);
 	close(pipe_fd[0]);
 	heredoc_prompt(delimeter, pipe_fd);
-	g_sigint = sig_bkp;
-	ft_signals(EXEC);
 }
 
 static void	heredoc_error(char *delimeter, t_sh *sh)
@@ -60,6 +54,7 @@ static int	parent_heredoc(t_sh *sh, int *pipe_fd, pid_t pid, char *delimeter)
 
 	close(pipe_fd[1]);
 	waitpid(pid, &sh->cl.exit_status, 0);
+	ft_signals(EXEC);
 	status = WEXITSTATUS(sh->cl.exit_status);
 	if (status == CTRLD)
 		heredoc_error(delimeter, sh);
@@ -76,15 +71,20 @@ int	heredoc(t_sh *sh, char *delimeter, int isbuiltin)
 	int		pipe_fd[2];
 	pid_t	pid;
 	int		status;
+	int		sig_bkp;
 
+	sig_bkp = g_sigint;
+	g_sigint = 0;
 	if (pipe(pipe_fd) < 0)
 		exit_error(strerror(errno));
+	ft_signals(UNSET);
 	pid = fork();
 	if (pid < 0)
 		exit_error(strerror(errno));
 	if (pid == 0)
 		execute_in_context(pipe_fd, delimeter);
 	status = parent_heredoc(sh, pipe_fd, pid, delimeter);
+	g_sigint = sig_bkp;
 	if (status == EXIT_SUCCESS)
 	{
 		if (isbuiltin)
